:construction: Fix DataManager implementation, and partially change model definition.
A fix was needed for the DataManager because it has some syntax errors and improperly
uses the nlohmann/json library.

Models that contain references to other models (e.g. the `prerequisites` member of
`Course`) now have the contents of the containers for references from objects to
pointers. This change from objects to pointers is necessary to prevent duplication,
and to have a more conceptually-correct code. References are supposed to point to
other objects. Using an object will violate this premise.