:sparkles: Add an additional ID to a Class object.
This is to allow for separation of individual class objects. The original
ID, which was used to identify which class objects are under which class,
has now been renamed to classID. This has also fixed a bug where only one
class object is being generated by the preprocessor  for a class.