:building_construction: Change preprocessor to give the generated data to the data manager.
The preprocessor is only there to generate data from some inputs. It is best that data is
stored in a central location for easier access. The data manager is there for centrally
storing data. Thus, we changed the preprocessor from keeping the data with itself to
saving the data to the data manager.