:sparkles: Add support for lab courses.
The preprocessor can now generate classes for lab units.

There was a bug prior to this commit that is related to generating classes for lab units. This bug
assigns a student to the same course multiple times. This effectively gives more slots to one
student than he/she needs in a class group, making the number of classes generated inadequate.